# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all trading accounts
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Add a trading account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ──────────────── TRADES ────────────────
...
  /education/sources:
    get:
      operationId: listEducationSources
      tags: [education]
      summary: List all knowledge base sources
      responses:
        "200":
          description: Education sources list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EducationSource"
    post:
      operationId: createEducationSource
      tags: [education]
      summary: Add a new knowledge source (book, video, playlist, raw text)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationSourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationSource"
  /education/sources/{id}/file:
    post:
      operationId: uploadSourceFile
      tags: [education]
      summary: Upload a PDF, EPUB, or supported book-page image for ingestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Supported extensions: PDF, EPUB, JPG, JPEG, PNG, WEBP, GIF, BMP, TIF, TIFF; declared media type must match the extension (application/octet-stream is permitted when the client omits a specific type)."
      responses:
        "200":
          description: File uploaded and ingestion triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
        "404":
          description: Source not found
        "400":
          description: Missing file, unsupported extension, or mismatched media type
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
        "413":
          description: File exceeds the 100 MB upload limit
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
  /education/sources/{id}/ingest:
    post:
      operationId: triggerEducationSourceIngest
      tags: [education]
      summary: Run or retry ingestion for an existing education source
      description: Ready sources return successfully without reprocessing; processing sources return 409 without starting a duplicate; pending and error sources are ingested/retried synchronously.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Ingestion completed, or the source was already ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400":
          description: Invalid source id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "409":
          description: Source is already processing; no duplicate ingestion was started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "500":
          description: Ingestion failed; error detail is persisted on the source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
  /education/sources/{id}:
    patch:
      operationId: updateEducationSource
      tags: [education]
      summary: Update ingestion status / metadata (called by C++ ingestion engine)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationSourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationSource"
        "404":
          description: Not found
    delete:
      operationId: deleteEducationSource
      tags: [education]
      summary: Remove a knowledge source from the system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ──────────────── BROKER CONNECTIONS ────────────────
...
  /brain/ingest/{id}:
    post:
      operationId: ingestSource
      tags: [brain]
      summary: Trigger ingestion of a specific education source
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ingestion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, broker, balance, equity, margin, currency, accountType, status, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        broker:      { type: string }
        balance:     { type: number }
        equity:      { type: number }
        margin:      { type: number }
        marginLevel: { type: ["number", "null"] }
        currency:    { type: string }
        accountType: { type: string, enum: [live, demo, prop] }
        status:      { type: string, enum: [active, inactive, suspended] }
        notes:       { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    AccountInput:
      type: object
      required: [name, broker, currency, accountType]
      properties:
        name:        { type: string, minLength: 1 }
        broker:      { type: string, minLength: 1 }
        balance:     { type: number, default: 0 }
        equity:      { type: number, default: 0 }
        margin:      { type: number, default: 0 }
        marginLevel: { type: number }
        currency:    { type: string, default: "USD" }
        accountType: { type: string, enum: [live, demo, prop] }
        status:      { type: string, enum: [active, inactive, suspended], default: active }
        notes:       { type: string }
    AccountUpdate:
      type: object
      properties:
        name:        { type: string }
        balance:     { type: number }
        equity:      { type: number }
        margin:      { type: number }
        marginLevel: { type: number }
        status:      { type: string, enum: [active, inactive, suspended] }
        notes:       { type: string }
...
    EducationSource:
      type: object
      required: [id, kind, title, status, chunksCount, vectorCount, createdAt]
      properties:
        id:                 { type: integer }
        kind:               { type: string, enum: [book, video, playlist, text] }
        title:              { type: string }
        author:             { type: ["string", "null"] }
        sourceUrl:          { type: ["string", "null"] }
        contentText:        { type: ["string", "null"] }
        status:             { type: string, enum: [pending, processing, ready, partial, error] }
        chunksCount:        { type: integer }
        vectorCount:        { type: integer }
        transcriptPreview:  { type: ["string", "null"] }
        metadata:           { type: ["string", "null"] }
        errorMessage:       { type: ["string", "null"] }
        createdAt:          { type: string, format: date-time }
        updatedAt:          { type: string, format: date-time }
    EducationSourceInput:
      type: object
      required: [kind, title]
      properties:
        kind:        { type: string, enum: [book, video, playlist, text] }
        title:       { type: string, minLength: 1 }
        author:      { type: string }
        sourceUrl:   { type: string }
        contentText: { type: string }
        metadata:    { type: string }
    EducationSourceUpdate:
      type: object
      properties:
        status:            { type: string, enum: [pending, processing, ready, partial, error] }
        chunksCount:       { type: integer, minimum: 0 }
        vectorCount:       { type: integer, minimum: 0 }
        transcriptPreview: { type: string }
        metadata:          { type: string }
        errorMessage:      { type: string }

    # ──────────── Broker Connections ────────────
...
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        message: { type: string }
    IngestResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        sourceId: { type: integer }
        error: { type: string }
        status: { type: string, enum: [pending, processing, ready, partial, error] }
        message: { type: string }
...